==> HW8/0.split_dataset_to_train_val.ipynb <==
# Разбейте in_domain_train на train и val

Мы использовали набор данных [RuCola](https://github.com/RussianNLP/RuCoLA)

import pandas as pd
from sklearn.model_selection import train_test_split

def get_appropriate_columns(df):
    return df[['sentence', 'acceptable']]


df = get_appropriate_columns(pd.read_csv(r'in_domain_train.csv'))

df_train, df_val = train_test_split(
    df,
    test_size=0.12,
    stratify=df['acceptable'],
    random_state=1)

print('train:', len(df_train))
#print(df_train.head())
print('validation:', len(df_val))
#print(df_val.head())

df_train.to_csv('in_domain_train_subset.csv', index=False)
df_val.to_csv('in_domain_validation_subset.csv', index=False)

df = get_appropriate_columns(pd.read_csv(r'in_domain_dev.csv'))
df.to_csv('in_domain_test.csv', index=False)

Мы разбили файл in_domain_train.csv на:
- тренировочный набор данных in_domain_train_subset.csv;
- валидационный набор данных in_domain_validation_subset.csv.

В результирующих файлах сохранены только колонки sentence (текст) и acceptable (метка).
(Также, пересохранили in_domain_dev.csv в in_domain_test.csv.)
