==> HW8/2.few-or_zero-shot_GPT3.ipynb <==
# few-/zero-shot c GPT3

Скачаем и используем [ai-forever/rugpt3large_based_on_gpt2](https://huggingface.co/ai-forever/rugpt3large_based_on_gpt2/tree/main)

Идея классификации: используем loss модели для определения правильности построения фраз. Если фраза построена правильно, то у GPT не должно возникнуть затруднений в работе. Если фраза неправильная, то loss должен получиться выше.

Обучим гиперпараметр threshold (уровень разделения классов) по тренировочному набору данных, после этого проверим результаты на тестовом наборе данных.

Импорты

from sklearn.metrics import matthews_corrcoef
from torch.utils.data import TensorDataset
from tqdm import tqdm
from transformers import GPT2Tokenizer, GPT2LMHeadModel
import datetime
import numpy as np
import pandas as pd
import random
import time
import torch
import torch
import warnings
warnings.filterwarnings("ignore")

Параметры

PATH = r'H:\Инструменты\Windows\GPT or another LLM\ai-forever rugpt3large_based_on_gpt2 2020/'

# Set the seed value all over the place to make this reproducible.
seed_val = 42
random.seed(seed_val)
np.random.seed(seed_val)
torch.manual_seed(seed_val)
torch.cuda.manual_seed_all(seed_val)

## Необходимые определения

def calc_train_loss_threshold(model, tokenizer, df_train, n):
    positive_loss = []
    negative_loss = []

    for i in tqdm(range(min(len(df_train), n))):
        loss = inference(model, tokenizer, df_train.iloc[i].sentence)

        if df_train.iloc[i].acceptable == 1:
            positive_loss.append(loss)
        else:
            negative_loss.append(loss)

    # Здесь можно улучшить, если ввести доп.гиперпараметр, чтобы задавать: к какому классу ближе порог.
    # Сейчас усреднение делается ровно по среднему между объектами обоих классов.
    return (sum(negative_loss) / len(negative_loss) + sum(positive_loss) / len(positive_loss)) / 2.


def inference(model, tokenizer, phrase):
    with torch.no_grad():
        encodings = tokenizer(phrase, return_tensors='pt')
        input_ids = encodings.input_ids
        outputs = model(input_ids=input_ids, labels=input_ids)
        loss = outputs.loss

    return loss


def test(model, tokenizer, threshold, phrase, verbose = True):
    if verbose:
        print(phrase)

    loss = inference(model, tokenizer, phrase)

    if loss < threshold:
      return 1
    else:
      return 0


def get_test_predictions(model, tokenizer, threshold, df_test, n):
    predictions = []
    labels = []
    for i in tqdm(range(min(len(df_test), n))):
        acceptable_pred = test(model, tokenizer, threshold, df_test.iloc[i].sentence, verbose=False)
        predictions.append(acceptable_pred)
        labels.append(df_test.iloc[i].acceptable)

    return predictions, labels


def get_matthews_corrcoef_score(predictions, true_labels):
    # Calculate the MCC
    mcc = matthews_corrcoef(true_labels, predictions)
    return mcc

## Загрузим наборы данных

df_train = pd.read_csv('in_domain_train_subset.csv')
df_test = pd.read_csv('in_domain_test.csv')
print('Train dataset:\n', df_train.head())
print('Test dataset:\n', df_test.head())

## Загрузим tokenizer и модель

print('Loading the tokenizer...')
tokenizer = GPT2Tokenizer.from_pretrained(PATH)
print('Ok')

model = GPT2LMHeadModel.from_pretrained(PATH)

### Обучим гиперпараметр threshold на тренировочном наборе данных

threshold = calc_train_loss_threshold(model, tokenizer, df_train, 7000)

### Грубо проверим результат

print('threshold =', threshold)

phrases = ["Лесные запахи набегали волнами; в них смешалось дыхание можжевельника, вереска, брусники.",
           "Вчера в два часа магазин закрыт.",
           'Кракадыл и друсья',
           'Крокодил Гена и его друзья.'
]
expecteds = [1, 0, 0, 1]

for phrase, expected in zip(phrases, expecteds):
    print('Expected:\t', expected)
    print('Predict:\t', test(model, tokenizer, threshold, phrase))
    print()

Мы видим, что, в принципе, получается корректный результат.

### Проверим на тестовом наборе данных

predictions, true_labels = get_test_predictions(model, tokenizer, threshold, df_test, 1000)

mcc = get_matthews_corrcoef_score(predictions, true_labels)
print('Total MCC: %.3f' % mcc)

Выводы:
- на грубых примерах подход работает хорошо: когда фразы явно неправильные — модель это определяет;
- на сложных тонких примерах тестового набора данных результаты получились низкие, на сложном dataset с тонкими дефектами построения фраз лучше результат с Bert.
