==> HW8/3.RuT5.ipynb <==
# RuT5

Скачаем и используем [ai-forever/ruT5-base](https://huggingface.co/ai-forever/ruT5-base/tree/main)

«…Обучите и протестируйте модель RuT5 на данной задаче (пример finetun’а можете найти здесь https://github.com/RussianNLP/RuCoLA/blob/main/baselines/finetune_t5.py)…»

Импорты библиотек

import os
from functools import partial
from pathlib import Path

import pandas as pd
import numpy as np
from datasets import load_metric, Dataset, DatasetDict
from razdel import tokenize
from transformers import (
    DataCollatorForSeq2Seq,
    Seq2SeqTrainingArguments,
    Seq2SeqTrainer,
    T5Tokenizer,
    T5ForConditionalGeneration,
)

import warnings
warnings.filterwarnings("ignore")

Зададим глобальные параметры.

PATH = r'H:\Инструменты\Windows\GPT or another LLM\ruT5-base 2021'
ACCURACY = load_metric("accuracy", keep_in_memory=True)
MCC = load_metric("matthews_correlation", keep_in_memory=True)
POS_LABEL = "грамотно"
NEG_LABEL = "неграмотно"

CURRENT_DIR = Path('.')
DATA_DIR = CURRENT_DIR
TRAIN_FILE = DATA_DIR / "in_domain_train_subset.csv"
IN_DOMAIN_DEV_FILE = DATA_DIR / "in_domain_validation_subset.csv"
TEST_FILE = DATA_DIR / "in_domain_test.csv"

N_EPOCHS = 10
LR_VALUES = (1e-4,) # 1e-3)
DECAY_VALUES = (0,) # 1e-4)
BATCH_SIZE = 128

Необходимые определения.

def compute_metrics(p, tokenizer):
    string_preds = tokenizer.batch_decode(p.predictions, skip_special_tokens=True)
    int_preds = [1 if prediction == POS_LABEL else 0 for prediction in string_preds]

    labels = np.where(p.label_ids != -100, p.label_ids, tokenizer.pad_token_id)
    string_labels = tokenizer.batch_decode(labels, skip_special_tokens=True)
    int_labels = []

    for string_label in string_labels:
        if string_label == POS_LABEL:
            int_labels.append(1)
        elif string_label == NEG_LABEL or string_label == "":  # second case accounts for test data
            int_labels.append(0)
        else:
            raise ValueError()

    acc_result = ACCURACY.compute(predictions=int_preds, references=int_labels)
    mcc_result = MCC.compute(predictions=int_preds, references=int_labels)

    result = {"accuracy": acc_result["accuracy"], "mcc": mcc_result["matthews_correlation"]}
    return result


def preprocess_data(sentences, tokenizer):
    result = tokenizer(sentences["sentence"], padding=False)

    if "acceptable" in sentences:
        label_sequences = []
        for label in sentences['acceptable']:
            if label == 1:
                target_sequence = POS_LABEL
            elif label == 0:
                target_sequence = NEG_LABEL
            else:
                raise ValueError("Unknown class label")
            label_sequences.append(target_sequence)
    else:
        # a hack to avoid the "You have to specify either decoder_input_ids or decoder_inputs_embeds" error
        # for test data
        label_sequences = ["" for _ in sentences]

    result["labels"] = tokenizer(label_sequences, padding=False)["input_ids"]
    result["length"] = [len(list(tokenize(sentence))) for sentence in sentences['sentence']]
    return result


def read_splits(*, as_datasets):
    train_df, in_domain_dev_df, test_df = map(
        pd.read_csv, (TRAIN_FILE, IN_DOMAIN_DEV_FILE, TEST_FILE)
    )

    # concatenate datasets to get aggregate metrics
    dev_df = in_domain_dev_df #pd.concat((in_domain_dev_df, out_of_domain_dev_df))

    if as_datasets:
        train, dev, test = map(Dataset.from_pandas, (train_df, dev_df, test_df))
        return DatasetDict(train=train, dev=dev, test=test)
    else:
        return train_df, dev_df, test_df


def train(tokenizer, data_collator, tokenized_splits):
    # seed, lr, wd, bs
    dev_metrics_per_run = np.empty((len(LR_VALUES), len(DECAY_VALUES), 2))

    for i, learning_rate in enumerate(LR_VALUES):
        for j, weight_decay in enumerate(DECAY_VALUES):
            model = T5ForConditionalGeneration.from_pretrained(PATH)

            run_base_dir = f"ruT5_{learning_rate}_{weight_decay}_{BATCH_SIZE}"

            training_args = Seq2SeqTrainingArguments(
                output_dir=f"checkpoints/{run_base_dir}",
                overwrite_output_dir=True,
                evaluation_strategy="epoch",
                per_device_train_batch_size=BATCH_SIZE,
                per_device_eval_batch_size=BATCH_SIZE,
                learning_rate=learning_rate,
                weight_decay=weight_decay,
                num_train_epochs=N_EPOCHS,
                lr_scheduler_type="constant",
                save_strategy="epoch",
                save_total_limit=1,
                seed=1,
                fp16=True,
                dataloader_num_workers=4,
                group_by_length=True,
                report_to="none",
                load_best_model_at_end=True,
                metric_for_best_model="eval_mcc",
                optim="adafactor",
                predict_with_generate=True,
            )

            trainer = Seq2SeqTrainer(
                model=model,
                args=training_args,
                train_dataset=tokenized_splits["train"],
                eval_dataset=tokenized_splits["dev"],
                compute_metrics=partial(compute_metrics, tokenizer=tokenizer),
                tokenizer=tokenizer,
                data_collator=data_collator,
            )

            train_result = trainer.train()
            print(f"{run_base_dir}")
            print("train", train_result.metrics)

            os.makedirs(f"results/{run_base_dir}", exist_ok=True)

            dev_predictions = trainer.predict(
                test_dataset=tokenized_splits["dev"], metric_key_prefix="validation", max_length=10
            )
            print("dev", dev_predictions.metrics)
            dev_metrics_per_run[i, j] = (
                dev_predictions.metrics["validation_mcc"],
            )

            #rmtree(f"checkpoints/{run_base_dir}")
    
    os.makedirs("results_agg", exist_ok=True)
    np.save(f"results_agg/T5_dev.npy", dev_metrics_per_run)    

    return trainer

## Загрузим tokenizer

tokenizer = T5Tokenizer.from_pretrained("ai-forever/ruT5-base")
#tokenizer.save_pretrained(PATH)

## Загрузим и подготовим наборы данных

splits_data = read_splits(as_datasets=True)

tokenized_splits_data = splits_data.map(
    partial(preprocess_data, tokenizer=tokenizer),
    batched=True,
    remove_columns=["sentence"],
)

data_collator = DataCollatorForSeq2Seq(tokenizer, pad_to_multiple_of=8)

## Тренировка модели (fine-tune)

trainer = train(tokenizer, data_collator, tokenized_splits_data)

## Проверим результат на тестовом наборе данных

predictions = trainer.predict(test_dataset=tokenized_splits_data["test"], max_length=10)
print("test", predictions.metrics)

string_preds = tokenizer.batch_decode(predictions.predictions, skip_special_tokens=True)

int_preds = [1 if prediction == POS_LABEL else 0 for prediction in string_preds]
int_preds = np.asarray(int_preds)

np.save(f"results/preds.npy", int_preds)
