==> HW8/4.Recap.ipynb <==
# Выводы

Сравним результаты по выбранной метрике: коэффициент корреляции Мэтьюса (MCC).

Список в формате «Модель:  MCC».
- RuBert-base: 0.390
- RuGPT3large_based_on_GPT2: 0.072
- RuT5-base:  0.328

На задаче бинарной классификации текста со сложным анализом (на правильность предложений по правилам русского языка) лучше всего результат с трансформерной моделью типа Encoder: RuBert.
